+++ src/com/gerardng/networkchat/Login.java
@@ -109,5 +109,6 @@ public class Login extends JFrame {
 	private void login(String name, String address, int port) {
 		// TODO Auto-generated method stub
 		System.out.println(name + " " + address + " " + port);
+		new Client(name, address, port);
 	}
 }