+++ src/com/gerardng/networkchat/server/Server.java
@@ -1,0 +1,10 @@
+package com.gerardng.networkchat.server;
+
+public class Server {
+	private int port;
+	
+	public Server(int port) {
+		this.port = port;
+	}
+
+}