+++ src/com/gerardng/networkchat/server/ServerMain.java
@@ -1,0 +1,5 @@
+package com.gerardng.networkchat.server;
+
+public class ServerMain {
+
+}