+++ changelog.md
@@ -4,7 +4,14 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 - Project is still under construction
 
-## [0.1.0] - 2017-01-19
+## [0.1.1] - 2017-01-19
+### Added
+- Built text history and text message fields.
+- Function to output messages on console
+- UI changes
+- Socket and InetAddress to open connection
+
+## [0.1.0] - 2017-01-18
 ### Added
 - WindowBuilder dependency for Java
 - Login form